+++ firmware/src/mqtt_client.h
@@ -6,6 +6,7 @@
 #include "battery.h"
 #include "schedule.h"
 #include "mqtt_settings.h"
+#include "pump_current.h"
 
 // Pubblica lo stato periodicamente e implementa Home Assistant MQTT
 // Discovery: appena connesso al broker, pubblica la configurazione delle
@@ -18,7 +19,7 @@ class MqttClientWrapper {
   void begin(MqttSettings& settings, Pump& pump);
   void loop();  // gestisce riconnessione non bloccante + pubblicazione periodica
   bool connected();
-  void publishStatusIfDue(Pump& pump, Battery& battery, Schedule& schedule);
+  void publishStatusIfDue(Pump& pump, Battery& battery, Schedule& schedule, PumpCurrentMonitor& pumpCurrent);
 
   // Da chiamare dopo che l'utente ha salvato una nuova configurazione da web:
   // si disconnette e riparte con i nuovi parametri, senza bisogno di riflashare.