+++ firmware/src/webserver.h
@@ -5,11 +5,13 @@
 #include "schedule.h"
 #include "mqtt_settings.h"
 #include "mqtt_client.h"
+#include "pump_current.h"
 
 class WebServerApp {
  public:
   WebServerApp(AsyncWebServer& server, Pump& pump, Battery& battery, Schedule& schedule,
-               bool& mqttConnected, MqttSettings& mqttSettings, MqttClientWrapper& mqttClient);
+               bool& mqttConnected, MqttSettings& mqttSettings, MqttClientWrapper& mqttClient,
+               PumpCurrentMonitor& pumpCurrentMonitor);
   void begin();
 
  private:
@@ -20,6 +22,7 @@ class WebServerApp {
   bool& mqttConnected_;
   MqttSettings& mqttSettings_;
   MqttClientWrapper& mqttClient_;
+  PumpCurrentMonitor& pumpCurrentMonitor_;
 
   void handleStatus(AsyncWebServerRequest* request);
   void handleManualStart(AsyncWebServerRequest* request, JsonVariant& body);